Add files via upload
The primary objective of this project is to create a system that can distinguish between pictures of cats and dogs. Following analysis of the original picture, a prediction of the result is made. You can get the Dogs vs. Cats dataset from the Kaggle website. There are pictures of cats and dogs in the collection. The model is meant to acquire different characteristics that make cats and dogs unique. When the model has finished being trained, it will be able to tell pictures of cats and dogs apart.